assignment3a/requirements.txt
Flask==2.0.3
google-cloud-storage
google-cloud-datastore
